students/wth/experiment3/requirements.txt
numpy
TensorFlow-gpu
pillow
matplotlib
